added functionality for fetching all components for select updated docs
