+++ functions/src/routes/adminRoutes/index.ts
@@ -4,6 +4,7 @@ import technologyRoutes from "./technologyRoutes";
 import clientRoutes from "./clientRoutes";
 import projectRoutes from "./projectRoutes";
 import portfolioRoutes from "./portfolioRoutes";
+import componentRoutes from "./componentRoutes";
 
 const router: Router = express.Router();
 
@@ -12,5 +13,6 @@ router.use("/technology", technologyRoutes);
 router.use("/client", clientRoutes);
 router.use("/project", projectRoutes);
 router.use("/portfolio", portfolioRoutes);
+router.use("/component", componentRoutes);
 
 export default router;
