+++ public/admin_api_doc.dev.yml
@@ -323,6 +323,18 @@ components:
           type: string
           description: updated at ISO string
           example: "2019-10-13T07:47:09.376Z"
+    SelectObject:
+      type: object
+      properties:
+        id:
+          type: string
+          description: unique identifier for the item
+          example: 31765c4c-2606-4e1b-a613-7866212a86b6
+        name:
+          type: string
+          description: name of the item
+          example: Mobile
+
     ItemInsertUpdateResponse:
       type: object
       properties:
@@ -1416,6 +1428,30 @@ paths:
           $ref: "#/components/responses/UnauthorizedError"
         404:
           $ref: "#/components/responses/ComponentNotFoundError"
+  /component/select:
+    get:
+      tags:
+        - component
+      description: Fetches all components for display in a select. This endpoint is to be used only when displaying data for a select
+      summary: Fetches all components for display in a select
+      responses:
+        200:
+          description: Successful response
+          content:
+            "application/json":
+              schema:
+                type: object
+                properties:
+                  payload:
+                    type: object
+                    properties:
+                      results:
+                        type: array
+                        items:
+                          $ref: "#/components/schemas/SelectObject"
+
+        401:
+          $ref: "#/components/responses/UnauthorizedError"
   /portfolio:
     post:
       tags:
